requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
arxiv==2.1.3
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
boto3==1.34.140
botocore==1.34.140
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
executing==2.0.1
Faker==26.0.0
feedparser==6.0.11
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-text-splitters==0.2.2
langsmith==0.1.83
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.35.10
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
phidata==2.4.25
pillow==10.4.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydeck==0.8.1b0
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pypdf==4.0.1
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
setuptools==70.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
singlestoredb==1.4.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparams==6.0.1
stack-data==0.6.3
streamlit==1.36.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
wheel==0.43.0
yarl==1.9.4
